% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageCohort.R
\name{ageCohort}
\alias{ageCohort}
\title{Create cohorts based on patients birthday}
\usage{
ageCohort(
  cdm,
  name,
  age = 0,
  ageUnit = "years",
  cohortName = "birthday_{age}_{ageUnit}"
)
}
\arguments{
\item{cdm}{A cdm reference.}

\item{name}{Name of the new cohort table created in the cdm object.}

\item{age}{Age of entry and exit to the cohort. Multiple values can be
supplied.}

\item{ageUnit}{Unit for the age, it can either be \emph{years} (default), or
\emph{days}. Multiple values can be supplied. Its length must be 1 or the same
than \code{age}.}

\item{cohortName}{Names of the created cohort, you can provide glue patterns
like: \verb{cohort_\{age\}_\{ageUnit\}} or a vector with as many names as cohorts as
created.}
}
\value{
A cohort table object.
}
\description{
\code{ageCohort()} creates a cohort table based on patient bithday. The
cohort entry and exit date will be the \code{nth} bithday of the individual.
cohort entry ends.
}
\examples{
\donttest{
library(CohortConstructor)

cdm <- mockCohortConstructor()

cdm$cohort4 <- ageCohort(cdm = cdm, name = "cohort4", age = 2)

attrition(cdm$cohort4)

# Can also create multiple age cohorts

cdm$cohort5 <- ageCohort(
  cdm = cdm,
  name = "cohort5",
  age = c(15, 3),
  ageUnit = c("days", "years")
)

attrition(cdm$cohort5)
}
}
