% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{atFirstDoc}
\alias{atFirstDoc}
\title{Helper for consistent documentation of arguments in \code{atFirst}
functions.}
\arguments{
\item{atFirst}{If FALSE the requirement will be applied to all records, if
TRUE, it will only be required for the first entry of each subject.}
}
\description{
Helper for consistent documentation of arguments in \code{atFirst}
functions.
}
\keyword{internal}
