% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{baseCohortDoc}
\alias{baseCohortDoc}
\title{Helper for consistent documentation of \code{conceptCohort} and \code{measurementCohort}.}
\arguments{
\item{typeConceptId}{Type concept id used to subset tables. See:
https://athena.ohdsi.org/search-terms/terms?conceptClass=Type+Concept. If
\code{NULL} no filtering is applied.}

\item{useRecordsBeforeObservation}{If FALSE, only records that fall entirely
(start and end) within an observation period are used.
If TRUE, records that start before an observation period are included by
shifting their start date to the beginning of the individual's subsequent
observation period, and records that start or end outside an observation
period are trimmed so that cohort records fall entirely within the
corresponding observation period.}

\item{useSourceFields}{If TRUE, the source concept_id fields will also be
used when identifying relevant clinical records. If FALSE, only the standard
concept_id fields will be used.}

\item{subsetCohort}{A character refering to a cohort table containing
individuals for whom cohorts will be generated. Only individuals in this
table will appear in the generated cohort.}

\item{subsetCohortId}{Optional. Specifies cohort IDs from the \code{subsetCohort}
table to include. If none are provided, all cohorts from the \code{subsetCohort}
are included.}
}
\description{
Helper for consistent documentation of \code{conceptCohort} and \code{measurementCohort}.
}
\keyword{internal}
