% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmarkCohortConstructor}
\alias{benchmarkCohortConstructor}
\title{Run benchmark of CohortConstructor package}
\usage{
benchmarkCohortConstructor(
  cdm,
  runCIRCE = TRUE,
  runCohortConstructorDefinition = TRUE,
  runCohortConstructorDomain = TRUE,
  dropCohorts = TRUE
)
}
\arguments{
\item{cdm}{A cdm reference.}

\item{runCIRCE}{Whether to run cohorts from JSON definitions generated with
Atlas.}

\item{runCohortConstructorDefinition}{Whether to run the benchmark part where
cohorts are created with CohortConstructor by definition (one by one,
separately).}

\item{runCohortConstructorDomain}{Whether to run the benchmark part where
cohorts are created with CohortConstructor by domain (instantianting base
cohort all together, as a set).}

\item{dropCohorts}{Whether to drop cohorts created during benchmark.}
}
\description{
Run benchmark of CohortConstructor cohort instantiation time compared to
CIRCE from JSON. More information in the benchmarking vignette.
}
