% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeCohortName.R
\name{changeCohortName}
\alias{changeCohortName}
\title{Utility function to change the name of a cohort.}
\usage{
changeCohortName(cohort, cohortId, newCohortName)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{Vector identifying which cohorts to include
(cohort_definition_id or cohort_name).}

\item{newCohortName}{Character vector with same}
}
\value{
A cohort_table object.
}
\description{
Utility function to change the name of a cohort.
}
\examples{
\donttest{
library(CohortConstructor)

cdm <- mockCohortConstructor(nPerson = 100)

settings(cdm$cohort1)

cdm$cohort1 <- cdm$cohort1 |>
  changeCohortName(cohortId = 1, newCohortName = "new_name")

settings(cdm$cohort1)
}
}
