% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonCohorts.R
\name{codeFromJson}
\alias{codeFromJson}
\title{Extract the code to generate an \code{ATLAS} json cohort.}
\usage{
codeFromJson(jsonPath, cdm)
}
\arguments{
\item{jsonPath}{A path to a folder with json files.}

\item{cdm}{A cdm_reference object.}
}
\value{
A character vector with the code to create the json equivalent cohorts.
}
\description{
Extract the code to generate an \code{ATLAS} json cohort.
}
