% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonCohorts.R
\name{codelistFromJson}
\alias{codelistFromJson}
\title{Extract codelists from a set of json files.}
\usage{
codelistFromJson(jsonPath, cdm)
}
\arguments{
\item{jsonPath}{A path to a folder with json files.}

\item{cdm}{A cdm_reference object.}
}
\value{
A \code{codelist} object.
}
\description{
Extract codelists from a set of json files.
}
