% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstructor-package.R
\docType{package}
\name{CohortConstructor-package}
\alias{CohortConstructor}
\alias{CohortConstructor-package}
\title{CohortConstructor: Build and Manipulate Study Cohorts Using a Common Data Model}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create and manipulate study cohorts in data mapped to the Observational Medical Outcomes Partnership Common Data Model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortConstructor/}
  \item \url{https://github.com/OHDSI/CohortConstructor}
  \item Report bugs at \url{https://github.com/OHDSI/CohortConstructor/issues}
}

}
\author{
\strong{Maintainer}: Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID})

Authors:
\itemize{
  \item Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID})
  \item Martí Català \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})
  \item Nuria Mercade-Besora \email{nuria.mercadebesora@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0006-7948-3747}{ORCID})
  \item Marta Alcalde-Herraiz \email{marta.alcaldeherraiz@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0002-4405-1814}{ORCID})
  \item Mike Du \email{mike.du@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9517-8834}{ORCID})
  \item Yuchen Guo \email{yuchen.guo@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-0847-4855}{ORCID})
  \item Xihang Chen \email{xihang.chen@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0001-8112-8959}{ORCID})
  \item Kim Lopez-Guell \email{kim.lopez@spc.ox.ac.uk} (\href{https://orcid.org/0000-0002-8462-8668}{ORCID})
  \item Elin Rowlands \email{elin.rowlands@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0005-5166-0417}{ORCID})
}

}
\keyword{internal}
