% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{cohortDefinitionDoc}
\alias{cohortDefinitionDoc}
\title{Helper for consistent documentation of \code{cohortDefinition}.}
\arguments{
\item{cohortDefinition}{A \code{cohort_definition} object containing one or more
cohort definitions.}
}
\description{
Helper for consistent documentation of \code{cohortDefinition}.
}
\keyword{internal}
