% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{cohortIdModifyDoc}
\alias{cohortIdModifyDoc}
\title{Helper for consistent documentation of \code{cohortId}.}
\arguments{
\item{cohortId}{IDs of the cohorts to modify. If NULL, all cohorts will be
used; otherwise, only the specified cohorts will be modified, and the
rest will remain unchanged.}
}
\description{
Helper for consistent documentation of \code{cohortId}.
}
\keyword{internal}
