% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{cohortIdSubsetDoc}
\alias{cohortIdSubsetDoc}
\title{Helper for consistent documentation of \code{cohortId}.}
\arguments{
\item{cohortId}{IDs of the cohorts to include. Cohorts not included will be
removed from the cohort set.}
}
\description{
Helper for consistent documentation of \code{cohortId}.
}
\keyword{internal}
