% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{collapseDoc}
\alias{collapseDoc}
\title{Helper for consistent documentation of \code{collapse}.}
\arguments{
\item{collapse}{Whether to collapse the overlapping records (TRUE) or drop
the records that have an ongoing prior record.}
}
\description{
Helper for consistent documentation of \code{collapse}.
}
\keyword{internal}
