% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{columnDateDoc}
\alias{columnDateDoc}
\title{Helper for consistent documentation of \code{dateColumns} and \code{returnReason}.}
\arguments{
\item{dateColumns}{Character vector indicating date columns in the cohort
table to consider.}

\item{returnReason}{If TRUE it will return a column indicating which of the
\code{dateColumns} was used.}

\item{multipleReasons}{If TRUE multiple reasons can be returned (reasons will
appear in the order of \code{dateColumns} separated by \verb{; }), if FALSE only the
first reason based on the order of \code{dateColumns} will be returned.}

\item{keepDateColumns}{If TRUE the returned cohort will keep columns in
\code{dateColumns}.}
}
\description{
Helper for consistent documentation of \code{dateColumns} and \code{returnReason}.
}
\keyword{internal}
