% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptCohortSet.R
\name{conceptCohortSet}
\alias{conceptCohortSet}
\title{Generate a cohort based on a concept set. The current supported domains are:}
\usage{
conceptCohortSet(cdm, conceptSet = NULL, name = "cohort", verbose = TRUE)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{conceptSet}{A conceptSet, can either be a list of concepts, a codelist
or a conceptSetExpression (TO DO).}

\item{name}{Name of the cohort in the cdm object.}

\item{verbose}{Whether to print intermediate progress messages.}
}
\value{
A cohort_table object.
}
\description{
Generate a cohort based on a concept set. The current supported domains are:
}
