% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{conceptSetDoc}
\alias{conceptSetDoc}
\title{Helper for consistent documentation of \code{conceptSet}.}
\arguments{
\item{conceptSet}{A conceptSet, which can either be a codelist
or a conceptSetExpression.}
}
\description{
Helper for consistent documentation of \code{conceptSet}.
}
\keyword{internal}
