% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{daysDoc}
\alias{daysDoc}
\title{Helper for consistent documentation of \code{days}.}
\arguments{
\item{days}{Integer with the number of days to add or name of a column (that
must be numeric) to add.}
}
\description{
Helper for consistent documentation of \code{days}.
}
\keyword{internal}
