% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deathCohort.R
\name{deathCohort}
\alias{deathCohort}
\title{Create cohort based on the death table}
\usage{
deathCohort(cdm, name, subsetCohort = NULL, subsetCohortId = NULL)
}
\arguments{
\item{cdm}{A cdm reference.}

\item{name}{Name of the new cohort table created in the cdm object.}

\item{subsetCohort}{A character refering to a cohort table containing
individuals for whom cohorts will be generated. Only individuals in this
table will appear in the generated cohort.}

\item{subsetCohortId}{Optional. Specifies cohort IDs from the \code{subsetCohort}
table to include. If none are provided, all cohorts from the \code{subsetCohort}
are included.}
}
\value{
A cohort table with a death cohort in cdm
}
\description{
Create cohort based on the death table
}
