% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entryAtColumnDate.R
\name{entryAtFirstDate}
\alias{entryAtFirstDate}
\title{Update cohort start date to be the first date from of a set of column dates}
\usage{
entryAtFirstDate(
  cohort,
  dateColumns,
  cohortId = NULL,
  returnReason = FALSE,
  multipleReasons = FALSE,
  keepDateColumns = TRUE,
  name = tableName(cohort),
  .softValidation = FALSE
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{dateColumns}{Character vector indicating date columns in the cohort
table to consider.}

\item{cohortId}{Vector identifying which cohorts to modify
(cohort_definition_id or cohort_name). If NULL, all cohorts will be
used; otherwise, only the specified cohorts will be modified, and the
rest will remain unchanged.}

\item{returnReason}{If TRUE it will return a column indicating which of the
\code{dateColumns} was used.}

\item{multipleReasons}{If TRUE multiple reasons can be returned (reasons will
appear in the order of \code{dateColumns} separated by \verb{; }), if FALSE only the
first reason based on the order of \code{dateColumns} will be returned.}

\item{keepDateColumns}{If TRUE the returned cohort will keep columns in
\code{dateColumns}.}

\item{name}{Name of the new cohort table created in the cdm object.}

\item{.softValidation}{Whether to perform a soft validation of consistency.
If set to FALSE four additional checks will be performed: 1) a check that
cohort end date is not before cohort start date,  2) a check that there
are no missing values in required columns, 3) a check that cohort duration is
all within observation period, and 4) that there are no overlapping
cohort entries}
}
\value{
The cohort table.
}
\description{
\code{entryAtFirstDate()} resets cohort start date based on a set of specified
column dates. The first date that occurs is chosen.
}
\examples{
\donttest{
library(CohortConstructor)
library(PatientProfiles)
cdm <- mockCohortConstructor()

cdm$cohort1 <- cdm$cohort1 |>
  addTableIntersectDate(
    tableName = "drug_exposure",
    nameStyle = "prior_drug",
    order = "last",
    window = c(-Inf, 0)
  ) |>
  addPriorObservation(priorObservationType = "date", name = "cohort1")

cdm$cohort1 |>
  entryAtFirstDate(dateColumns = c("prior_drug", "prior_observation"))
}
}
