% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erafy.R
\name{erafy}
\alias{erafy}
\title{Collapse cohort entries using a certain gap to concatenate records.}
\usage{
erafy(cohort, cohortId = NULL, gap = 0, name = tableName(cohort))
}
\arguments{
\item{cohort}{A cohort table}

\item{cohortId}{The cohort definition ids to subset, if NULL all cohort
definition ids are used.}

\item{gap}{Number of days to use when merging cohort entries.}

\item{name}{Name of the cohort table.}
}
\value{
A cohort table
}
\description{
Collapse cohort entries using a certain gap to concatenate records.
}
