% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCohortDefinition.R
\name{exportCohortDefinition}
\alias{exportCohortDefinition}
\title{Title}
\usage{
exportCohortDefinition(x, path)
}
\arguments{
\item{x}{A cohort definition.}

\item{path}{An existing directory in which to write one JSON file per
cohort definition.}
}
\description{
Title
}
