% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{gapDoc}
\alias{gapDoc}
\title{Helper for consistent documentation of \code{gap}.}
\arguments{
\item{gap}{Number of days between two subsequent cohort entries to be merged
in a single cohort record.}
}
\description{
Helper for consistent documentation of \code{gap}.
}
\keyword{internal}
