% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateIntersectCohorts.R
\name{generateIntersectCohortSet}
\alias{generateIntersectCohortSet}
\title{Generate a combination cohort set between the intersection of different
cohorts.}
\usage{
generateIntersectCohortSet(
  cdm,
  name,
  targetCohortName,
  targetCohortId = NULL,
  mutuallyExclusive = FALSE,
  returnOnlyComb = FALSE
)
}
\arguments{
\item{cdm}{A cdm reference.}

\item{name}{Name of the new generated cohort.}

\item{targetCohortName}{Name of an existent cohort in the cdm to create the
combinations.}

\item{targetCohortId}{Ids to combine of the target cohort. If NULL all
cohort present in the table will be used.}

\item{mutuallyExclusive}{Whether the generated cohorts are mutually
exclusive or not.}

\item{returnOnlyComb}{Whether to only get the combination cohort back}
}
\value{
The cdm object with the new generated cohort set
}
\description{
Generate a combination cohort set between the intersection of different
cohorts.
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()

cdm <- generateIntersectCohortSet(
  cdm = cdm,
  name = "cohort3",
  targetCohortName = "cohort2"
)

cdm$cohort3

CDMConnector::settings(cdm$cohort3)

}
}
