% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMatchedCohortSet.R
\name{generateMatchedCohortSet}
\alias{generateMatchedCohortSet}
\title{Generate a new cohort matched cohort from a preexisting target cohort. The
new cohort will contain individuals not included in the target cohort with
same year of birth (matchYearOfBirth = TRUE) and same sex (matchSex = TRUE).}
\usage{
generateMatchedCohortSet(
  cdm,
  name,
  targetCohortName,
  targetCohortId = NULL,
  matchSex = TRUE,
  matchYearOfBirth = TRUE,
  ratio = 1
)
}
\arguments{
\item{cdm}{A cdm reference object.}

\item{name}{Name of the new generated cohort set.}

\item{targetCohortName}{Name of the target cohort to match.}

\item{targetCohortId}{Cohort definition id to match from the target cohort.
If NULL all the cohort definition id present in the target cohort will be
matched.}

\item{matchSex}{Whether to match in sex.}

\item{matchYearOfBirth}{Whether to match in year of birth.}

\item{ratio}{Number of allowed matches per individual in the target cohort.}
}
\value{
A cdm reference object that contains the new generated cohort set.
}
\description{
Generate a new cohort matched cohort from a preexisting target cohort. The
new cohort will contain individuals not included in the target cohort with
same year of birth (matchYearOfBirth = TRUE) and same sex (matchSex = TRUE).
}
