% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateStrataCohortSet.R
\name{generateStrataCohortSet}
\alias{generateStrataCohortSet}
\title{Generate a stratified cohort set from an existent cohort.}
\usage{
generateStrataCohortSet(
  cdm,
  name,
  targetCohortName,
  targetCohortId = NULL,
  strataColumn = character()
)
}
\arguments{
\item{cdm}{A cdm reference.}

\item{name}{Name of the new generated cohort.}

\item{targetCohortName}{Name of an existent cohort in the cdm to create the
combinations.}

\item{targetCohortId}{Ids to stratify of the target cohort. If NULL all
cohort present in the table will be used.}

\item{strataColumn}{Columns in targetCohortName that you want to use to
stratificatify.}
}
\value{
The cdm object with the new generated cohort set
}
\description{
Generate a stratified cohort set from an existent cohort.
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()

cdm$cohort2 <- cdm$cohort2 |>
  addSex()

cdm <- generateStrataCohortSet(
  cdm = cdm,
  name = "cohort3",
  targetCohortName = "cohort2",
  strataColumn = "sex"
)

cdm$cohort3

settings(cdm$cohort3)

}
}
