% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectCohorts.R
\name{getIdentifier}
\alias{getIdentifier}
\title{Get random identifiers not present in a table based on a prefix.}
\usage{
getIdentifier(x, len = 1, prefix = "", nchar = 5)
}
\arguments{
\item{x}{Table.}

\item{len}{Number of identifiers.}

\item{prefix}{Character vector with the prefix of the identifiers.}

\item{nchar}{Number of random characters added to the prefix.}
}
\value{
Character vector of identifiers not present in x.
}
\description{
Get random identifiers not present in a table based on a prefix.
}
