% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCohortDefinition.R
\name{importCohortDefinition}
\alias{importCohortDefinition}
\title{Title}
\usage{
importCohortDefinition(path, recursive = FALSE)
}
\arguments{
\item{path}{A path, URL, or directory containing JSON cohort definitions.}

\item{recursive}{Whether JSON files should be searched for recursively when
\code{path} is a directory.}
}
\description{
Title
}
