% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectCohorts.R
\name{joinOverlap}
\alias{joinOverlap}
\title{Join overlapping periods in single periods.}
\usage{
joinOverlap(
  x,
  start = "cohort_start_date",
  end = "cohort_end_date",
  by = c("cohort_definition_id", "subject_id"),
  gap = 0
)
}
\arguments{
\item{x}{Table in the cdm.}

\item{start}{Column that indicates the start of periods.}

\item{end}{Column that indicates the end of periods.}

\item{by}{Variables to group by.}

\item{gap}{Distance between exposures to consider that they overlap.}
}
\value{
Table in the cdm with start, end and by as columns. Periods are not
going to overlap between each other.
}
\description{
Join overlapping periods in single periods.
}
