% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonCohorts.R
\name{jsonCohort}
\alias{jsonCohort}
\title{Create cohort from an 'ATLAS' defined json file.}
\usage{
jsonCohort(cdm, jsonPath, name)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{jsonPath}{Path to json files.}

\item{name}{Name of the cohort to create.}
}
\value{
A cdm_reference object with the cohort created. # DO we want to output cohort
}
\description{
Create cohort from an 'ATLAS' defined json file.
}
