% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{keepOriginalCohortsDoc}
\alias{keepOriginalCohortsDoc}
\title{Helper for consistent documentation of \code{keepOriginalCohorts}.}
\arguments{
\item{keepOriginalCohorts}{If TRUE the original cohorts will be return
together with the new ones. If FALSE only the new cohort will be returned.}
}
\description{
Helper for consistent documentation of \code{keepOriginalCohorts}.
}
\keyword{internal}
