% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockCohortConstructor.R
\name{mockCohortConstructor}
\alias{mockCohortConstructor}
\title{Function to create a mock cdm reference for CohortConstructor}
\usage{
mockCohortConstructor(source = "local", seed = 1)
}
\arguments{
\item{source}{Source for the mock cdm, it can either be 'local' or 'duckdb'.}

\item{seed}{An optional integer used to set the seed for random number
generation, ensuring reproducibility of the generated data. If provided,
this seed allows the function to produce consistent results each time.
If 'NULL', the seed is not set, which can lead to different outputs on
each run.}
}
\value{
cdm object
}
\description{
\code{mockCohortConstructor()} creates an example dataset that can be used for
demonstrating and testing the package
}
\examples{
\donttest{
library(CohortConstructor)
cdm <- mockCohortConstructor()

cdm
}
}
