% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{nameDoc}
\alias{nameDoc}
\title{Helper for consistent documentation of \code{name}.}
\arguments{
\item{name}{Name of the new cohort table created in the cdm object.}
}
\description{
Helper for consistent documentation of \code{name}.
}
\keyword{internal}
