% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padCohortEnd.R
\name{padCohortEnd}
\alias{padCohortEnd}
\title{Add days to cohort end}
\usage{
padCohortEnd(cohort, days, cohortId = NULL, name = tableName(cohort))
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{days}{Number of day to add to the cohort end date.}

\item{cohortId}{Vector identifying which cohorts to modify
(cohort_definition_id or cohort_name). If NULL, all cohorts will be
used; otherwise, only the specified cohorts will be modified, and the
rest will remain unchanged.}

\item{name}{Name of the new cohort table created in the cdm object.}
}
\value{
Cohort table
}
\description{
\code{padCohortEnd()} Adds (or subtracts) a certain number of days to the cohort
end date. Note:
\itemize{
\item If the days added means that cohort end would be after observation
period end date, then observation period end date will be used for cohort
exit.
\item If the days added means that cohort exit would be after the next cohort
start then these overlapping cohort entries will be collapsed.
\item If days subtracted means that cohort end would be before cohort start then
the cohort entry will be dropped.
}
}
\examples{
\donttest{
library(CohortConstructor)
cdm <- mockCohortConstructor()
# add 10 days to each cohort exit
cdm$cohort1 |>
  padCohortEnd(days = 10)
}
}
