% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{padObservationDoc}
\alias{padObservationDoc}
\title{Helper for consistent documentation of \code{padObservation}.}
\arguments{
\item{padObservation}{Whether to pad observations if they are outside
observation_period (TRUE) or drop the records if they are outside
observation_period (FALSE)}
}
\description{
Helper for consistent documentation of \code{padObservation}.
}
\keyword{internal}
