% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireDemographics.R
\name{requireAge}
\alias{requireAge}
\title{Restrict cohort on age}
\usage{
requireAge(
  cohort,
  cohortId = NULL,
  ageRange = list(c(0, 150)),
  indexDate = "cohort_start_date",
  name = omopgenerics::tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{Vector of cohort definition ids to include. If NULL, all
cohort definition ids will be used.}

\item{ageRange}{A list of minimum and maximum age.}

\item{indexDate}{Variable in cohort that contains the date to compute the
demographics characteristics on which to restrict on.}

\item{name}{Name of the new cohort with the age requirement.}

\item{keep}{If TRUE cohorts used to apply restrictions will be kept as they
were, otherwise they will be dropped from the cohort table.}
}
\value{
The cohort table with only records for individuals satisfying the
age requirement
}
\description{
Restrict cohort on age
}
\examples{
library(DrugUtilisation)
library(CohortConstructor)
cdm <- mockDrugUtilisation(numberIndividuals = 100)
cdm$cohort1 \%>\%
  requireAge(indexDate = "cohort_start_date",
             ageRange = list(c(18, 65)))
}
