% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireCohortIntersectFlag.R
\name{requireCohortIntersectFlag}
\alias{requireCohortIntersectFlag}
\title{Require cohort subjects are present in another cohort}
\usage{
requireCohortIntersectFlag(
  x,
  targetCohortTable,
  targetCohortId = NULL,
  indexDate = "cohort_start_date",
  targetStartDate = "cohort_start_date",
  targetEndDate = "cohort_end_date",
  window = list(c(0, Inf)),
  negate = FALSE
)
}
\arguments{
\item{x}{Cohort table}

\item{targetCohortTable}{name of the cohort that we want to check for overlap}

\item{targetCohortId}{vector of cohort definition ids to include}

\item{indexDate}{Variable in x that contains the date to compute the
intersection.}

\item{targetStartDate}{date of reference in cohort table, either for start
(in overlap) or on its own (for incidence)}

\item{targetEndDate}{date of reference in cohort table, either for end
(overlap) or NULL (if incidence)}

\item{window}{window to consider events over}

\item{negate}{If set as TRUE, criteria will be applied as exclusion
rather than inclusion (i.e. require absence in another cohort)}
}
\value{
Cohort table with only those in the other cohort kept
}
\description{
Require cohort subjects are present in another cohort
}
\examples{
library(DrugUtilisation)
library(CohortConstructor)
cdm <- mockDrugUtilisation(numberIndividuals = 100)
cdm$cohort1 \%>\%
  requireCohortIntersectFlag(targetCohortTable = "cohort2",
                             targetCohortId = 1,
                             indexDate = "cohort_start_date",
                             window = c(-Inf, 0))
}
