% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireConceptIntersectFlag.R
\name{requireConceptIntersectCount}
\alias{requireConceptIntersectCount}
\title{Require cohort subjects to have an amount of events of a concept list.}
\usage{
requireConceptIntersectCount(
  cohort,
  conceptSet,
  window,
  counts,
  requirementType = "exactly",
  cohortId = NULL,
  indexDate = "cohort_start_date",
  targetStartDate = "event_start_date",
  targetEndDate = "event_end_date",
  censorDate = NULL,
  name = tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{conceptSet}{Concept set list.}

\item{window}{Window to consider events over.}

\item{counts}{Number of intersections to require.}

\item{requirementType}{Can take the values: "exactly", "at_least" or
"at_most". It indicates whether the number of intersections required should
match the counts ("excatly"), or if the counts refer to the minimum
("at_least") or maximum ("at_most") number of intersections required.}

\item{cohortId}{IDs of the cohorts to modify. If NULL, all cohorts will be
used; otherwise, only the specified cohorts will be modified, and the
rest will remain unchanged..}

\item{indexDate}{Variable in x that contains the date to compute the
intersection.}

\item{targetStartDate}{Date of reference in cohort table, either for start
(in overlap) or on its own (for incidence).}

\item{targetEndDate}{Date of reference in cohort table, either for end
(overlap) or NULL (if incidence).}

\item{censorDate}{Whether to censor overlap events at a specific date or a
column date of x.}

\item{name}{Name of the new cohort with the future observation restriction.}
}
\value{
Cohort table.
}
\description{
\code{requireConceptIntersectCount()} filters a cohort table based on a requirement
that an individual has a certain amount of events related to a concept
list in some time window around an index date.
}
\examples{
\donttest{
library(CohortConstructor)
cdm <- mockCohortConstructor(conditionOccurrence = TRUE)
cdm$cohort2 <-  requireConceptIntersectCount(
  cohort = cdm$cohort1,
  conceptSet = list(a = 1),
  window = c(-Inf, 0),
  counts = 2,
  name = "cohort2")
  }
}
