% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireCurrentEvent.R
\name{requireCurrentEvent}
\alias{requireCurrentEvent}
\title{Filter clinical tables to keep only current events}
\usage{
requireCurrentEvent(
  cdm,
  tables = c("condition_occurrence", "procedure_occurrence", "observation")
)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{tables}{A character vector specifying the names of the tables to filter.
Default is \code{c("condition_occurrence", "procedure_occurrence", "observation")}.}
}
\value{
A modified cdm_reference object where the specified tables are filtered.
}
\description{
\code{requireCurrentEvent()} filters clinical tables in a CDM object, keeping only
records where the date field and the date part of the datetime field are equal.
This is used to distinguish confirmed "current events" from historical or
discordant events.
}
\examples{
\donttest{
library(CohortConstructor)
library(omopgenerics)
cdm <- mockCohortConstructor()
cdm <- requireCurrentEvent(cdm)
}
}
