% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireDeathFlag.R
\name{requireDeathFlag}
\alias{requireDeathFlag}
\title{Require cohort subjects' death at a certain time window}
\usage{
requireDeathFlag(
  cohort,
  window,
  cohortId = NULL,
  indexDate = "cohort_start_date",
  censorDate = NULL,
  negate = FALSE,
  name = tableName(cohort),
  .softValidation = FALSE
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{window}{Window to consider events over.}

\item{cohortId}{Vector of cohort definition ids to include. If NULL, all
cohort definition ids will be used.}

\item{indexDate}{Variable in x that contains the date to compute the
intersection.}

\item{censorDate}{Whether to censor overlap events at a specific date or a
column date of x.}

\item{negate}{If set as TRUE, criteria will be applied as exclusion
rather than inclusion (i.e. require absence in another cohort).}

\item{name}{Name of the new cohort with the future observation restriction.}

\item{.softValidation}{Whether to perform a soft validation of consistency.
If set to FALSE four additional checks will be performed: 1) check that
cohort end date is not before cohort start date, 2) check that there are no
missing values in required columns, 3) check that cohort duration is all
within observation period, and 4) check that there are no overlapping cohort
entries.}
}
\value{
Cohort table with only those with a death event kept (or without
if negate = TRUE)
}
\description{
Require cohort subjects' death at a certain time window
}
\examples{
library(CDMConnector)
library(CohortConstructor)
cdm <- mockCohortConstructor(death = TRUE)
cdm$cohort1 <- cdm$cohort1 |> requireDeathFlag(window = list(c(0, Inf)))
attrition(cdm$cohort1)
}
