% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{requireDemographicsDoc}
\alias{requireDemographicsDoc}
\title{Helper for consistent documentation of arguments in \code{requireDemographics}.}
\arguments{
\item{ageRange}{A list of vectors specifying minimum and maximum age.}

\item{sex}{Can be "Both", "Male" or "Female".}

\item{minPriorObservation}{A minimum number of continuous prior observation
days in the database.}

\item{minFutureObservation}{A minimum number of continuous future observation
days in the database.}

\item{indexDate}{Variable in cohort that contains the date to compute the
demographics characteristics on which to restrict on.}

\item{requirementInteractions}{If TRUE, cohorts will be created for
all combinations of ageGroup, sex, and daysPriorObservation. If FALSE, only
the first value specified for the other factors will be used. Consequently,
order of values matters when requirementInteractions is FALSE.}
}
\description{
Helper for consistent documentation of arguments in \code{requireDemographics}.
}
\keyword{internal}
