% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{requireFullContributionDoc}
\alias{requireFullContributionDoc}
\title{Helper for consistent documentation of \code{requireFullContribution}.}
\arguments{
\item{requireFullContribution}{Whether to require individuals to contribute all
required days. If TRUE, those individuals for which adding days would take
them out of observation will be dropped. If FALSE, days will only be added
up to the day when the individual leaves observation.}
}
\description{
Helper for consistent documentation of \code{requireFullContribution}.
}
\keyword{internal}
