% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireIsFirstEntry.R
\name{requireIsLastEntry}
\alias{requireIsLastEntry}
\title{Restrict cohort to last entry by index date}
\usage{
requireIsLastEntry(
  cohort,
  cohortId = NULL,
  indexDate = "cohort_start_date",
  name = tableName(cohort),
  .softValidation = FALSE
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{Vector of cohort definition ids to include. If NULL, all
cohort definition ids will be used.}

\item{indexDate}{Column name in cohort that contains the date to restrict on.}

\item{name}{Name of the new cohort with the restriction.}

\item{.softValidation}{Whether to perform a soft validation of consistency.
If set to FALSE four additional checks will be performed: 1) check that
cohort end date is not before cohort start date, 2) check that there are no
missing values in required columns, 3) check that cohort duration is all
within observation period, and 4) check that there are no overlapping cohort
entries.}
}
\value{
A cohort table in a cdm reference.
}
\description{
Restrict cohort to last entry by index date
}
\examples{
\donttest{
library(CohortConstructor)
cdm <- mockCohortConstructor()
cdm$cohort1 <- requireIsLastEntry(cdm$cohort1)
}

}
