% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireMinimumCohortCount.R
\name{requireMinimumCohortCount}
\alias{requireMinimumCohortCount}
\title{Generate a cohort table using a subset of cohorts from another table.}
\usage{
requireMinimumCohortCount(
  cohort,
  cohortId = NULL,
  minCohortCount = 5,
  name = tableName(cohort)
)

requireMinimumCohortCount(
  cohort,
  cohortId = NULL,
  minCohortCount = 5,
  name = tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{IDs of the cohorts to apply minimum cell count requirement.
If NULL it will be applied to all cohorts.}

\item{minCohortCount}{The minimum count of sbjects for a cohort to be
included.}

\item{name}{Name of the new cohort with the demographic requirements.}
}
\value{
Cohort table

Cohort table
}
\description{
\code{requireMinimumCohortCount()} filters an existing cohort table, keeping only
records from cohorts with a minimum number of individuals

\code{requireMinimumCohortCount()} filters an existing cohort table, keeping only
records from cohorts with a minimum number of individuals
}
\examples{
\donttest{
library(CohortConstructor)

cdm <- mockCohortConstructor(nPerson = 100)

cdm$cohort1 |>
requireMinimumCohortCount(minCohortCount = 5)
}
\donttest{
library(CohortConstructor)

cdm <- mockCohortConstructor(nPerson = 100)

cdm$cohort1 |>
requireMinimumCohortCount(minCohortCount = 5)
}
}
