% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireDemographics.R
\name{requirePriorObservation}
\alias{requirePriorObservation}
\title{Restrict cohort on prior observation}
\usage{
requirePriorObservation(
  cohort,
  indexDate = "cohort_start_date",
  minPriorObservation = 0
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference}

\item{indexDate}{Variable in cohort that contains the date to compute the
demographics characteristics on which to restrict on.}

\item{minPriorObservation}{A mimimum number of prior observation days in
the database.}
}
\description{
Restrict cohort on prior observation
}
