% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireTableIntersectFlag.R
\name{requireTableIntersectCount}
\alias{requireTableIntersectCount}
\title{Require cohort subjects have a certain amount of records in another clinical
table}
\usage{
requireTableIntersectCount(
  cohort,
  tableName,
  window,
  counts,
  requirementType = "exactly",
  cohortId = NULL,
  indexDate = "cohort_start_date",
  targetStartDate = startDateColumn(tableName),
  targetEndDate = endDateColumn(tableName),
  censorDate = NULL,
  name = tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{tableName}{Name of the table to check for intersect.}

\item{window}{Window to consider events over.}

\item{counts}{Number of intersections to require.}

\item{requirementType}{Can take the values: "exactly", "at_least" or
"at_most". It indicates whether the number of intersections required should
match the counts ("excatly"), or if the counts refer to the minimum
("at_least") or maximum ("at_most") number of intersections required.}

\item{cohortId}{IDs of the cohorts to modify. If NULL, all cohorts will be
used; otherwise, only the specified cohorts will be modified, and the
rest will remain unchanged.}

\item{indexDate}{Variable in x that contains the date to compute the
intersection.}

\item{targetStartDate}{Date of reference in cohort table, either for start
(in overlap) or on its own (for incidence).}

\item{targetEndDate}{Date of reference in cohort table, either for end
(overlap) or NULL (if incidence).}

\item{censorDate}{Whether to censor overlap events at a specific date or a
column date of x.}

\item{name}{Name of the new cohort with the future observation restriction.}
}
\value{
Cohort table.
}
\description{
\code{requireTableIntersectCount()} filters a cohort table based on a requirement
that an individual has a certain amount of records in
a clinical table in some time window around an index date.
}
\examples{
\donttest{
library(CohortConstructor)
cdm <- mockCohortConstructor(drugExposure = TRUE)
cdm$cohort1 |>
  requireTableIntersectCount(tableName = "drug_exposure",
                            indexDate = "cohort_start_date",
                            window = c(-Inf, 0),
                            counts = 2)
}
}
