% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictToFirstEntry.R
\name{restrictToFirstEntry}
\alias{restrictToFirstEntry}
\title{Restrict cohort to first entry by index date}
\usage{
restrictToFirstEntry(cohort, indexDate = "cohort_start_date")
}
\arguments{
\item{cohort}{A cohort table in a cdm reference}

\item{indexDate}{indexDate Variable in cohort that contains the date to
restrict on}
}
\value{
a cohort table in a cdm reference
}
\description{
Restrict cohort to first entry by index date
}
