% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{simpleReasonDoc}
\alias{simpleReasonDoc}
\title{Helper for consistent documentation of \code{reason}.}
\arguments{
\item{reason}{A character vector to populate the reason in the attrition.}
}
\description{
Helper for consistent documentation of \code{reason}.
}
\keyword{internal}
