% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{softValidationDoc}
\alias{softValidationDoc}
\title{Helper for consistent documentation of \code{.softValidation}.}
\arguments{
\item{.softValidation}{Whether to perform a soft validation of consistency.
If set to FALSE four additional checks will be performed: 1) a check that
cohort end date is not before cohort start date,  2) a check that there
are no missing values in required columns, 3) a check that cohort duration is
all within observation period, and 4) that there are no overlapping
cohort entries}
}
\description{
Helper for consistent documentation of \code{.softValidation}.
}
\keyword{internal}
