% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectCohorts.R
\name{splitOverlap}
\alias{splitOverlap}
\title{To split overlaping periods in non overlaping period.}
\usage{
splitOverlap(
  x,
  start = "cohort_start_date",
  end = "cohort_end_date",
  by = c("cohort_definition_id", "subject_id")
)
}
\arguments{
\item{x}{Table in the cdm.}

\item{start}{Column that indicates the start of periods.}

\item{end}{Column that indicates the end of periods.}

\item{by}{Variables to group by.}
}
\value{
Table in the cdm with start, end and by as columns. Periods are not
going to overlap between each other.
}
\description{
To split overlaping periods in non overlaping period.
}
