% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseCohortOverlap.R
\name{summariseCohortOverlap}
\alias{summariseCohortOverlap}
\title{Summarise cohort overlap}
\usage{
summariseCohortOverlap(
  cohort,
  restrictToFirstEntry = TRUE,
  timing = c("min", "q25", "median", "q75", "max")
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference}

\item{restrictToFirstEntry}{If TRUE only an individual's first entry per
cohort will be considered. If FALSE all entries per individual will be
considered}

\item{timing}{Summary statistics for timing. If NULL, timings between cohort
entries will not be considered}
}
\value{
A summarised result
}
\description{
Summarise cohort overlap
}
