% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireDateRange.R
\name{trimToDateRange}
\alias{trimToDateRange}
\title{Trim cohort dates to be within a date range}
\usage{
trimToDateRange(
  cohort,
  startDate = "cohort_start_date",
  endDate = "cohort_end_date",
  dateRange = as.Date(c(NA, NA))
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference}

\item{startDate}{Variable with earliest date}

\item{endDate}{Variable with latest date}

\item{dateRange}{A window of time during which the index date must have
been observed}
}
\description{
Trim cohort dates to be within a date range
}
