% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionCohort.R
\name{unionCohort}
\alias{unionCohort}
\title{Generate cohort from the union of different cohorts}
\usage{
unionCohort(
  cohort,
  cohortId = NULL,
  gap = 0,
  cohortName = NULL,
  name = omopgenerics::tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{Vector of cohort definition ids to include. If NULL, all
cohort definition ids will be used.}

\item{gap}{Number of days between two subsequent cohort entries of a subject
that will be merged in a single cohort entry}

\item{cohortName}{Name of the returned cohort. If NULL, the cohort name will
be created by collapsing the individual cohort names, separated by "_".}

\item{name}{Name of the new cohort table.}
}
\value{
The cdm object with the new generated cohort set
}
\description{
Generate cohort from the union of different cohorts
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()
cdm$cohort2 <- cdm$cohort2 |> unionCohort()
CDMConnector::settings(cdm$cohort2)

}
}
