% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{windowDoc}
\alias{windowDoc}
\title{Helper for consistent documentation of \code{window}.}
\arguments{
\item{window}{A list of vectors specifying minimum and maximum days from
\code{indexDate} to consider events over.}
}
\description{
Helper for consistent documentation of \code{window}.
}
\keyword{internal}
