% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{baseCohortDoc}
\alias{baseCohortDoc}
\title{Helper for consistent documentation of \code{conceptCohort} and \code{measurementCohort}.}
\arguments{
\item{useRecordsBeforeObservation}{If FALSE, only records in observation will
be used. If TRUE, records before the start of observation period will be
considered, with cohort start date set as the start date of the
individuals next observation period (as cohort records must be within
observation).}

\item{useSourceFields}{If TRUE, the source concept_id fields will also be
used when identifying relevant clinical records. If FALSE, only the standard
concept_id fields will be used.}

\item{subsetCohort}{A character refering to a cohort table containing
individuals for whom cohorts will be generated. Only individuals in this
table will appear in the generated cohort.}

\item{subsetCohortId}{Optional. Specifies cohort IDs from the \code{subsetCohort}
table to include. If none are provided, all cohorts from the \code{subsetCohort}
are included.}
}
\description{
Helper for consistent documentation of \code{conceptCohort} and \code{measurementCohort}.
}
\keyword{internal}
