% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstructor-package.R
\docType{package}
\name{CohortConstructor-package}
\alias{CohortConstructor}
\alias{CohortConstructor-package}
\title{CohortConstructor: Generate Cohorts of Individuals in the OMOP Common Data Model}
\description{
This package aims to provide functionalities to create cohorts in the OMOP Common Data Model.
}
\author{
\strong{Maintainer}: Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID})

Authors:
\itemize{
  \item Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})
}

}
\keyword{internal}
