% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstructor-package.R
\docType{package}
\name{CohortConstructor-package}
\alias{CohortConstructor}
\alias{CohortConstructor-package}
\title{CohortConstructor: Work With Cohorts Using a Common Data Model}
\description{
This package aims to provide functionalities to manipulate and evaluate cohorts in data mapped to the Observational Medical Outcomes Partnership Common Data Model.
}
\author{
\strong{Maintainer}: Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID})

Authors:
\itemize{
  \item Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})
  \item Marta Alcalde \email{marta.alcaldeherraiz@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0002-4405-1814}{ORCID})
  \item Yuchen Guo \email{yuchen.guo@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-0847-4855}{ORCID})
}

Other contributors:
\itemize{
  \item Nuria Mercade-Besora \email{nuria.mercadebesora@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0006-7948-3747}{ORCID}) [contributor]
}

}
\keyword{internal}
