% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseCohorts.R
\name{collapseCohorts}
\alias{collapseCohorts}
\title{Collapse cohort entries using a certain gap to concatenate records.}
\usage{
collapseCohorts(cohort, cohortId = NULL, gap = 0, name = tableName(cohort))
}
\arguments{
\item{cohort}{A cohort table}

\item{cohortId}{IDs of the cohorts to modify. If NULL, all cohorts will be
used; otherwise, only the specified cohorts will be modified, and the
rest will remain unchanged.}

\item{gap}{Number of days to use when merging cohort entries.}

\item{name}{Name of the cohort table.}
}
\value{
A cohort table
}
\description{
\code{collapseCohorts()} concatenates cohort records, allowing for some number
of days between one finishing and the next starting.
}
