% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptCohort.R
\name{conceptCohort}
\alias{conceptCohort}
\title{Create cohorts based on a concept set}
\usage{
conceptCohort(cdm, conceptSet, name, .softValidation = FALSE)
}
\arguments{
\item{cdm}{A cdm reference.}

\item{conceptSet}{A conceptSet, which can either be a codelist, a codelist
or a conceptSetExpression.}

\item{name}{Name of the cohort in the cdm object.}

\item{.softValidation}{Whether to perform a soft validation of consistency.
If set to FALSE four additional checks will be performed: 1) check that
cohort end date is not before cohort start date, 2) check that there are no
missing values in required columns, 3) check that cohort duration is all
within observation period, and 4) check that there are no overlapping cohort
entries.}
}
\value{
A cohort table
}
\description{
Create cohorts based on a concept set
}
\examples{
library(CohortConstructor)

cdm <- mockCohortConstructor(conditionOccurrence = TRUE)

cohort <- conceptCohort(cdm = cdm, conceptSet = list(a = 1), name = "cohort")

cohort |> attrition()

}
