% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptCohort.R
\name{conceptCohort}
\alias{conceptCohort}
\title{Generate a cohort based on a concept set. The current supported domains are:}
\usage{
conceptCohort(cdm, conceptSet, name)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{conceptSet}{A conceptSet, can either be a list of concepts, a codelist
or a conceptSetExpression (TO DO).}

\item{name}{Name of the cohort in the cdm object.}
}
\value{
A cohort_table object.
}
\description{
Generate a cohort based on a concept set. The current supported domains are:
}
