% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographicsCohort.R
\name{demographicsCohort}
\alias{demographicsCohort}
\title{Create the observation period cohort}
\usage{
demographicsCohort(
  cdm,
  name,
  ageRange = NULL,
  sex = NULL,
  minPriorObservation = NULL,
  minFutureObservation = NULL
)
}
\arguments{
\item{cdm}{A cdm_reference.}

\item{name}{Name of the new cohort_table object.}

\item{ageRange}{A list of minimum and maximum age.}

\item{sex}{Can be "Both", "Male" or "Female". If one of the latter, only
those with that sex will be included.}

\item{minPriorObservation}{A minimum number of prior observation days in
the database.}

\item{minFutureObservation}{A minimum number of future observation days in
the database.}
}
\value{
The cohort with the observation period
}
\description{
Create the observation period cohort
}
