% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erafy.R
\name{erafy}
\alias{erafy}
\title{Collapse a cohort using a certain gap to join records.}
\usage{
erafy(cohort, cohortId = NULL, gap = 0, name = tableName(cohort))
}
\arguments{
\item{cohort}{A cohort_table object.}

\item{cohortId}{The cohort definition ids to subset, if NULL all cohort
definition ids are used.}

\item{gap}{number of days to join consecutive records.}

\item{name}{Name of the resultant cohort.}
}
\value{
A cohort_table object.
}
\description{
Collapse a cohort using a certain gap to join records.
}
