% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortExit.R
\name{exitAtDeath}
\alias{exitAtDeath}
\title{Set cohort end date to end of observation}
\usage{
exitAtDeath(
  cohort,
  cohortId = NULL,
  requireDeath = FALSE,
  name = omopgenerics::tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{Vector of cohort definition ids to include. If NULL, all
cohort definition ids will be used.}

\item{requireDeath}{If TRUE, subjects without a death record will be dropped,
while if FALSE their end date will be left as is.}

\item{name}{Name of the new cohort with the restriction.}
}
\value{
The cohort table.
}
\description{
This functions changes cohort end date to the end date of the observation
period corresponding to the cohort entry. In the case were this generates
overlapping records in the cohort, overlapping entries will be merged.
}
\examples{
library(PatientProfiles)
library(CohortConstructor)
cdm <- mockPatientProfiles()
cdm$cohort1 \%>\% exitAtDeath()
}
