% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCohortDefinition.R
\name{instantiateCohortDefinition}
\alias{instantiateCohortDefinition}
\title{Instantiate a cohort from a cohort definition}
\usage{
instantiateCohortDefinition(cohortDefinition, cdm, name, conceptSet = NULL)
}
\arguments{
\item{cohortDefinition}{A \code{cohort_definition} object containing one or more
cohort definitions.}

\item{cdm}{A cdm reference.}

\item{name}{Name of the output cohort table in the \code{cdm}. All cohort
definitions in \code{cohortDefinition} are instantiated and bound into this
table.}

\item{conceptSet}{A conceptSet, which can either be a codelist
or a conceptSetExpression.}
}
\value{
A cohort table containing the instantiated cohort.
}
\description{
\code{instantiateCohortDefinition()} evaluates all stored cohort-definition
pipelines against an OMOP CDM, resolves their cohort dependencies, binds
the resulting cohort tables, and returns the combined cohort table.
}
\examples{
\donttest{
library(CohortConstructor)
cdm <- mockCohortConstructor()
definition <- cohortDefinitionFromCode(
  'cdm$cohort <- conceptCohort(cdm = cdm, name = "cohort", '
  'conceptSet = codelist["condition"])'
)
instantiateCohortDefinition(
  cohortDefinition = definition,
  cdm = cdm,
  name = "cohort",
  conceptSet = list(condition = 201826L)
)
}
}
