% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectCohorts.R
\name{intersectCohorts}
\alias{intersectCohorts}
\title{Generate a combination cohort set between the intersection of different
cohorts.}
\usage{
intersectCohorts(
  cohort,
  cohortId = NULL,
  gap = 0,
  mutuallyExclusive = FALSE,
  keepOriginalCohorts = FALSE,
  name = tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{IDs of the cohorts to include. Cohorts not included will be
removed from the cohort set.}

\item{gap}{Number of days between two subsequent cohort entries to be merged
in a single cohort record.}

\item{mutuallyExclusive}{Whether the generated cohorts are mutually
exclusive or not.}

\item{keepOriginalCohorts}{If TRUE the original cohorts and the newly
created intersection cohort will be returned. If FALSE only the new cohort
will be returned.}

\item{name}{Name of the new cohort table created in the cdm object.}
}
\value{
A cohort table.
}
\description{
\code{intersectCohorts()} combines different cohort entries, with those records
that overlap combined and kept. Cohort entries are when an individual was in
\emph{both} of the cohorts.
}
\examples{
\donttest{
library(CohortConstructor)

cdm <- mockCohortConstructor(nPerson = 100)

cdm$cohort3 <- intersectCohorts(
  cohort = cdm$cohort2,
  name = "cohort3",
)

settings(cdm$cohort3)

}
}
