% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectCohorts.R
\name{intersectCohorts}
\alias{intersectCohorts}
\title{Generate a combination cohort set between the intersection of different
cohorts.}
\usage{
intersectCohorts(
  cohort,
  cohortId = NULL,
  gap = 0,
  mutuallyExclusive = FALSE,
  returnOnlyComb = FALSE,
  name = tableName(cohort),
  .softValidation = FALSE
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{Vector of cohort definition ids to include. If NULL, all
cohort definition ids will be used.}

\item{gap}{Number of days between two subsequent cohort entries to be merged
in a single cohort record.}

\item{mutuallyExclusive}{Whether the generated cohorts are mutually
exclusive or not.}

\item{returnOnlyComb}{Whether to only get the combination cohort back}

\item{name}{Name of the new cohort with the demographic requirements.}

\item{.softValidation}{Whether to perform a soft validation of consistency.
If set to FALSE four additional checks will be performed: 1) check that
cohort end date is not before cohort start date, 2) check that there are no
missing values in required columns, 3) check that cohort duration is all
within observation period, and 4) check that there are no overlapping cohort
entries.}
}
\value{
A cohort table.
}
\description{
Generate a combination cohort set between the intersection of different
cohorts.
}
\examples{
\donttest{
library(CohortConstructor)

cdm <- mockCohortConstructor(nPerson = 100)

cdm$cohort3 <- intersectCohorts(
  cohort = cdm$cohort2,
  name = "cohort3",
)

settings(cdm$cohort3)

}
}
