% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectCohorts.R
\name{intersectCohorts}
\alias{intersectCohorts}
\title{Generate a combination cohort set between the intersection of different
cohorts.}
\usage{
intersectCohorts(
  cohort,
  cohortId = NULL,
  gap = 0,
  mutuallyExclusive = FALSE,
  returnOnlyComb = FALSE,
  name = omopgenerics::tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{Vector of cohort definition ids to include. If NULL, all
cohort definition ids will be used.}

\item{gap}{Number of days between two subsequent cohort entries to be merged
in a single cohort record.}

\item{mutuallyExclusive}{Whether the generated cohorts are mutually
exclusive or not.}

\item{returnOnlyComb}{Whether to only get the combination cohort back}

\item{name}{Name of the new cohort with the demographic requirements.}
}
\value{
The cdm object with the new generated cohort set
}
\description{
Generate a combination cohort set between the intersection of different
cohorts.
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()

cdm$cohort3 <- intersectCohorts(
  cohort = cdm$cohort1,
  name = "cohort3",
)

cdm$cohort3

CDMConnector::settings(cdm$cohort3)

}
}
