% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchCohorts.R
\name{matchCohorts}
\alias{matchCohorts}
\title{Generate a new cohort matched cohort from a preexisting target cohort. The
new cohort will contain individuals not included in the target cohort with
same year of birth (matchYearOfBirth = TRUE) and same sex (matchSex = TRUE).}
\usage{
matchCohorts(
  cohort,
  cohortId = NULL,
  matchSex = TRUE,
  matchYearOfBirth = TRUE,
  ratio = 1,
  name = omopgenerics::tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{Cohort definition id to match from the target cohort.
If NULL all the cohort definition id present in the target cohort will be
matched.}

\item{matchSex}{Whether to match in sex.}

\item{matchYearOfBirth}{Whether to match in year of birth.}

\item{ratio}{Number of allowed matches per individual in the target cohort.}

\item{name}{Name of the new generated cohort set.}
}
\value{
A cohort table.
}
\description{
Generate a new cohort matched cohort from a preexisting target cohort. The
new cohort will contain individuals not included in the target cohort with
same year of birth (matchYearOfBirth = TRUE) and same sex (matchSex = TRUE).
}
\examples{
library(PatientProfiles)
library(CohortConstructor)
library(dplyr)
cdm <- mockPatientProfiles(
  person = tibble(
    person_id = 1:3,
    gender_concept_id = c(8532, 8532, 8507),
    year_of_birth = 1992,
    month_of_birth = 1,
    day_of_birth = 1,
    race_concept_id = 0,
    ethnicity_concept_id = 0
  ),
  observation_period = tibble(
    observation_period_id = 1:3,
      person_id = 1:3,
      observation_period_start_date = as.Date("1950-01-01"),
      observation_period_end_date = as.Date("2023-01-01"),
      period_type_concept_id = 0
  )
)
cdm$new_matched_cohort <- cdm$cohort2 \%>\%
  matchCohorts(
    name = "new_matched_cohort",
    cohortId = 2,
    matchSex = TRUE,
    matchYearOfBirth = TRUE,
    ratio = 1)
cdm$new_matched_cohort

}
