% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurementCohort.R
\name{measurementCohort}
\alias{measurementCohort}
\title{Create cohorts based on a concept set from the measurement table}
\usage{
measurementCohort(
  cdm,
  conceptSet,
  name,
  valueAsConcept = NULL,
  valueAsNumber = NULL
)
}
\arguments{
\item{cdm}{A cdm reference.}

\item{conceptSet}{A conceptSet, which can either be a codelist
or a conceptSetExpression.}

\item{name}{Name of the cohort in the cdm object.}

\item{valueAsConcept}{A vector of cohort IDs used to filter measurements.
Only measurements with these values in the \code{value_as_concept_id} column of
the measurement table will be included. If NULL all entries independently of
their value as concept will be considered.}

\item{valueAsNumber}{A named list indicating the range of values and the unit
they correspond to, as follows:
list("unit_concept_id" = c(rangeValue1, rangeValue2)). If NULL, all entries
independently of their value as number will be included.}
}
\value{
A cohort table
}
\description{
This function creates cohorts from concepts in the measurement table. Value
of the target measurements in \code{conceptSet} can be specified through
\code{valueAsConcept} and \code{valueAsNumber}. If both NULL, conceptSets will not be
restricted to any value, if one of them is not NULL, value restrictions will
be based on the non NULL entry, and if both are specified, concepts
fulfilling one OR the other will be included.
}
\examples{
library(CohortConstructor)

cdm <- mockCohortConstructor(conditionOccurrence = TRUE)

cohort <- conceptCohort(cdm = cdm, conceptSet = list(a = 1), name = "cohort")

cohort |> attrition()

}
