% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockCohortConstructor.R
\name{mockCohortConstructor}
\alias{mockCohortConstructor}
\title{Function to create a mock cdm reference for CohortConstructor}
\usage{
mockCohortConstructor(
  nPerson = 10,
  conceptTable = NULL,
  tables = NULL,
  conceptId = NULL,
  conceptIdClass = NULL,
  drugExposure = F,
  conditionOccurrence = F,
  death = F,
  con = DBI::dbConnect(duckdb::duckdb())
)
}
\arguments{
\item{nPerson}{number of person in the cdm}

\item{conceptTable}{user defined concept table}

\item{tables}{list of tables to include in the cdm}

\item{conceptId}{list of concept id}

\item{conceptIdClass}{the domain class of the conceptId}

\item{drugExposure}{T/F include drug exposure table in the cdm}

\item{conditionOccurrence}{T/F include condition occurrence in the cdm}

\item{death}{T/F include death table in the cdm}

\item{con}{db connection detail for copy to databases}
}
\value{
cdm object
}
\description{
Function to create a mock cdm reference for CohortConstructor
}
\examples{
library(CohortConstructor)

cdm <- mockCohortConstructor()

cdm

}
