% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockCohortConstructor.R
\name{mockCohortConstructor}
\alias{mockCohortConstructor}
\title{Function to create a mock cdm reference for CohortConstructor}
\usage{
mockCohortConstructor(source = "local")
}
\arguments{
\item{source}{Source for the mock cdm, it can either be 'local' or 'duckdb'.}
}
\value{
cdm object
}
\description{
\code{mockCohortConstructor()} creates an example dataset that can be used for
demonstrating and testing the package
}
\examples{
\donttest{
library(CohortConstructor)
if(isTRUE(omock::isMockDatasetDownloaded("GiBleed"))){
cdm <- mockCohortConstructor()

cdm
}
}
}
