% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireDemographics.R
\name{requireAge}
\alias{requireAge}
\title{Restrict cohort on age}
\usage{
requireAge(cohort, indexDate = "cohort_start_date", ageRange = list(c(0, 150)))
}
\arguments{
\item{cohort}{A cohort table in a cdm reference}

\item{indexDate}{Variable in cohort that contains the date to compute the
demographics characteristics on which to restrict on.}

\item{ageRange}{A list of minimum and maximum age}
}
\description{
Restrict cohort on age
}
