% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireCohortIntersectFlag.R
\name{requireCohortIntersectFlag}
\alias{requireCohortIntersectFlag}
\title{Require cohort subjects are present in another cohort}
\usage{
requireCohortIntersectFlag(
  x,
  targetCohortTable,
  targetCohortId = NULL,
  indexDate = "cohort_start_date",
  targetStartDate = "cohort_start_date",
  targetEndDate = "cohort_end_date",
  window = list(c(0, Inf))
)
}
\arguments{
\item{x}{Cohort table}

\item{targetCohortTable}{name of the cohort that we want to check for overlap}

\item{targetCohortId}{vector of cohort definition ids to include}

\item{indexDate}{Variable in x that contains the date to compute the
intersection.}

\item{targetStartDate}{date of reference in cohort table, either for start
(in overlap) or on its own (for incidence)}

\item{targetEndDate}{date of reference in cohort table, either for end
(overlap) or NULL (if incidence)}

\item{window}{window to consider events of}
}
\value{
Cohort table with only those in the other cohort kept
}
\description{
Require cohort subjects are present in another cohort
}
