% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireConceptIntersectFlag.R
\name{requireConceptIntersectFlag}
\alias{requireConceptIntersectFlag}
\title{Require cohort subjects to have events of a concept list}
\usage{
requireConceptIntersectFlag(
  cohort,
  cohortId = NULL,
  conceptSet,
  indexDate = "cohort_start_date",
  targetStartDate = "event_start_date",
  targetEndDate = "event_end_date",
  censorDate = NULL,
  window = list(c(0, Inf)),
  negate = FALSE,
  name = omopgenerics::tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{Vector of cohort definition ids to include. If NULL, all
cohort definition ids will be used.}

\item{conceptSet}{Concept set list.}

\item{indexDate}{Variable in x that contains the date to compute the
intersection.}

\item{targetStartDate}{Date of reference in cohort table, either for start
(in overlap) or on its own (for incidence).}

\item{targetEndDate}{Date of reference in cohort table, either for end
(overlap) or NULL (if incidence).}

\item{censorDate}{Whether to censor overlap events at a specific date or a
column date of x.}

\item{window}{Window to consider events over.}

\item{negate}{If set as TRUE, criteria will be applied as exclusion
rather than inclusion (i.e. require absence in another cohort).}

\item{name}{Name of the new cohort with the future observation restriction.}
}
\value{
Cohort table with only those  with the events in the concept list
kept (or those without the event if nagate = TRUE)
}
\description{
Require cohort subjects to have events of a concept list
}
\examples{
library(PatientProfiles)
library(CohortConstructor)
cdm <- mockPatientProfiles()
cdm <- CDMConnector::insertTable(cdm, name = "concept",
                                 table = dplyr::tibble(
                                   "concept_id" = 1,
                                   "concept_name" = "my concept",
                                   "domain_id" = "Drug",
                                   "vocabulary_id" = NA,
                                   "concept_class_id" = NA,
                                   "concept_code" = NA,
                                   "valid_start_date" = NA,
                                   "valid_end_date" = NA
                                  ))
cdm$cohort2 <-  requireConceptIntersectFlag(
  cohort = cdm$cohort1,
  conceptSet = list(a = 1),
  window = c(-Inf, 0),
  name = "cohort2")
}
