% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireDeathFlag.R
\name{requireDeathFlag}
\alias{requireDeathFlag}
\title{Require cohort subjects' death at a certain time window}
\usage{
requireDeathFlag(
  x,
  indexDate = "cohort_start_date",
  censorDate = NULL,
  window = list(c(0, Inf)),
  negate = FALSE,
  name = omopgenerics::tableName(x)
)
}
\arguments{
\item{x}{Cohort table.}

\item{indexDate}{Variable in x that contains the date to compute the
intersection.}

\item{censorDate}{Whether to censor overlap events at a specific date or a
column date of x.}

\item{window}{Window to consider events over.}

\item{negate}{If set as TRUE, criteria will be applied as exclusion
rather than inclusion (i.e. require absence in another cohort).}

\item{name}{Name of the new cohort with the future observation restriction.}
}
\value{
Cohort table with only those in the other cohort kept
}
\description{
Require cohort subjects' death at a certain time window
}
\examples{
library(PatientProfiles)
library(CDMConnector)
library(CohortConstructor)
cdm <- mockPatientProfiles()
cdm <- insertTable(cdm, "death",
                   table = dplyr::tibble(
                     person_id = 1,
                     death_date = as.Date("2020-05-01"),
                     death_type_concept_id = NA))
cdm$cohort1 <- cdm$cohort1 \%>\% requireDeathFlag()
attrition(cdm$cohort1)
}
