% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireDemographics.R
\name{requireDemographics}
\alias{requireDemographics}
\title{Restrict cohort on patient demographics}
\usage{
requireDemographics(
  cohort,
  indexDate = "cohort_start_date",
  ageRange = list(c(0, 150)),
  sex = c("Both"),
  minPriorObservation = 0,
  minFutureObservation = 0
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference}

\item{indexDate}{Variable in cohort that contains the date to compute the
demographics characteristics on which to restrict on.}

\item{ageRange}{A list of minimum and maximum age}

\item{sex}{Can be "Both", "Male" or "Female". If one of the latter, only
those with that sex will be included.}

\item{minPriorObservation}{A mimimum number of prior observation days in
the database.}

\item{minFutureObservation}{A minimum number of future observation days in
the database.}
}
\value{
The cohort table with only records for individuals satisfying the
demographic requirements
}
\description{
Restrict cohort on patient demographics
}
\examples{
library(DrugUtilisation)
library(CohortConstructor)
cdm <- mockDrugUtilisation(numberIndividuals = 100)
cdm$cohort1 \%>\%
  requireDemographics(indexDate = "cohort_start_date",
                      ageRange = list(c(18, 65)),
                      sex = "Female",
                      minPriorObservation = 365)

}
