% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireDemographics.R
\name{requireFutureObservation}
\alias{requireFutureObservation}
\title{Restrict cohort on future observation}
\usage{
requireFutureObservation(
  cohort,
  indexDate = "cohort_start_date",
  minFutureObservation = 0
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference}

\item{indexDate}{Variable in cohort that contains the date to compute the
demographics characteristics on which to restrict on.}

\item{minFutureObservation}{A minimum number of future observation days in
the database.}
}
\value{
The cohort table with only records for individuals satisfying the
future observation requirement
}
\description{
Restrict cohort on future observation
}
\examples{
library(DrugUtilisation)
library(CohortConstructor)
cdm <- mockDrugUtilisation(numberIndividuals = 100)
cdm$cohort1 \%>\%
  requireFutureObservation(indexDate = "cohort_start_date",
                           minFutureObservation = 30)
}
