% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireDateRange.R
\name{requireInDateRange}
\alias{requireInDateRange}
\title{Require that an index date is within a date range}
\usage{
requireInDateRange(
  cohort,
  dateRange,
  cohortId = NULL,
  indexDate = "cohort_start_date",
  name = tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{dateRange}{A window of time during which the index date must have
been observed.}

\item{cohortId}{Vector of cohort definition ids to include. If NULL, all
cohort definition ids will be used.}

\item{indexDate}{Variable in cohort that contains the index date of interest}

\item{name}{Name of the new cohort with the restriction.}
}
\value{
The cohort table with any cohort entries outside of the date range
dropped
}
\description{
Require that an index date is within a date range
}
\examples{
library(CohortConstructor)

cdm <- mockCohortConstructor(nPerson = 100)
cdm$cohort1 |>
  requireInDateRange(indexDate = "cohort_start_date",
                     dateRange = as.Date(c("2010-01-01", "2019-01-01")))
}
