% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{requireIntersectDoc}
\alias{requireIntersectDoc}
\title{Helper for consistent documentation of arguments in \code{requireIntersect}
functions.}
\arguments{
\item{indexDate}{Name of the column in the cohort that contains the date to
compute the intersection.}

\item{intersections}{A range indicating number of intersections for
criteria to be fulfilled. If a single number is passed, the number of
intersections must match this.}

\item{targetStartDate}{Start date of reference in cohort table.}

\item{targetEndDate}{End date of reference in cohort table. If NULL,
incidence of target event in the window will be considered as intersection,
otherwise prevalence of that event will be used as intersection (overlap
between cohort and event).}

\item{censorDate}{Whether to censor overlap events at a specific date or a
column date of the cohort.}

\item{targetCohortTable}{Name of the cohort that we want to check for
intersect.}

\item{targetCohortId}{Vector of cohort definition ids to include.}

\item{tableName}{Name of the table to check for intersect.}

\item{inObservation}{If TRUE only records inside an observation period will
be considered.}
}
\description{
Helper for consistent documentation of arguments in \code{requireIntersect}
functions.
}
\keyword{internal}
