% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireIsEntry.R
\name{requireIsEntry}
\alias{requireIsEntry}
\title{Restrict cohort to specific entry}
\usage{
requireIsEntry(cohort, entryRange, cohortId = NULL, name = tableName(cohort))
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{entryRange}{Range for entries to include.}

\item{cohortId}{Vector identifying which cohorts to modify
(cohort_definition_id or cohort_name). If NULL, all cohorts will be
used; otherwise, only the specified cohorts will be modified, and the
rest will remain unchanged.}

\item{name}{Name of the new cohort table created in the cdm object.}
}
\value{
A cohort table in a cdm reference.
}
\description{
\code{requireIsFirstEntry()} filters cohort records, keeping only the first
cohort entry per person.
}
\examples{
\donttest{
library(CohortConstructor)
cdm <- mockCohortConstructor()
cdm$cohort1 <- requireIsEntry(cdm$cohort1, c(1, Inf))
}

}
