% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireIsLastEntry.R
\name{requireIsFirstEntry}
\alias{requireIsFirstEntry}
\title{Restrict cohort to first entry by index date}
\usage{
requireIsFirstEntry(
  cohort,
  cohortId = NULL,
  indexDate = "cohort_start_date",
  name = omopgenerics::tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{Vector of cohort definition ids to include. If NULL, all
cohort definition ids will be used.}

\item{indexDate}{Column name in cohort that contains the date to restrict on.}

\item{name}{Name of the new cohort with the restriction.}
}
\value{
A cohort table in a cdm reference.
}
\description{
Restrict cohort to first entry by index date
}
\examples{
\donttest{
library(CohortConstructor)
library(PatientProfiles)
cdm <- mockPatientProfiles()
cdm$cohort1 <- requireIsFirstEntry(cdm$cohort1)
}

}
