% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireMinCohortCount.R
\name{requireMinCohortCount}
\alias{requireMinCohortCount}
\title{Filter cohorts to keep only records for those with a minimum amount of subjects}
\usage{
requireMinCohortCount(
  cohort,
  cohortId = NULL,
  minCohortCount = 5,
  name = tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{IDs of the cohorts to modify. If NULL, all cohorts will be
used; otherwise, only the specified cohorts will be modified, and the
rest will remain unchanged.}

\item{minCohortCount}{The minimum count of sbjects for a cohort to be
included.}

\item{name}{Name of the new cohort table created in the cdm object.}
}
\value{
Cohort table
}
\description{
\code{requireMinCohortCount()} filters an existing cohort table, keeping only
records from cohorts with a minimum number of individuals
}
\examples{
\donttest{
library(CohortConstructor)

cdm <- mockCohortConstructor(nPerson = 100)

cdm$cohort1 |>
requireMinCohortCount(minCohortCount = 5)
}
}
