% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireDemographics.R
\name{requirePriorObservation}
\alias{requirePriorObservation}
\title{Restrict cohort on prior observation}
\usage{
requirePriorObservation(
  cohort,
  minPriorObservation,
  cohortId = NULL,
  indexDate = "cohort_start_date",
  name = omopgenerics::tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{minPriorObservation}{A minimum number of prior observation days in
the database.}

\item{cohortId}{Vector of cohort definition ids to include. If NULL, all
cohort definition ids will be used.}

\item{indexDate}{Variable in cohort that contains the date to compute the
demographics characteristics on which to restrict on.}

\item{name}{Name of the new cohort with the prior observation restriction.}
}
\value{
The cohort table with only records for individuals satisfying the
prior observation requirement
}
\description{
Restrict cohort on prior observation
}
\examples{
library(DrugUtilisation)
library(CohortConstructor)
cdm <- mockDrugUtilisation(numberIndividuals = 100)
cdm$cohort1 \%>\%
  requirePriorObservation(indexDate = "cohort_start_date",
                          minPriorObservation = 365)
}
