% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireDemographics.R
\name{requireSex}
\alias{requireSex}
\title{Restrict cohort on sex}
\usage{
requireSex(cohort, sex = c("Both"))
}
\arguments{
\item{cohort}{A cohort table in a cdm reference}

\item{sex}{Can be "Both", "Male" or "Female". If one of the latter, only
those with that sex will be included.}
}
\value{
The cohort table with only records for individuals satisfying the
sex requirement
}
\description{
Restrict cohort on sex
}
\examples{
library(DrugUtilisation)
library(CohortConstructor)
cdm <- mockDrugUtilisation(numberIndividuals = 100)
cdm$cohort1 \%>\%
  requireSex(sex = "Female")
}
