% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireDemographics.R
\name{requireSex}
\alias{requireSex}
\title{Restrict cohort on sex}
\usage{
requireSex(
  cohort,
  cohortId = NULL,
  sex = c("Both"),
  name = omopgenerics::tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{Vector of cohort definition ids to include. If NULL, all
cohort definition ids will be used.}

\item{sex}{Can be "Both", "Male" or "Female". If one of the latter, only
those with that sex will be included.}

\item{name}{Name of the new cohort with the sex requirements.}
}
\value{
The cohort table with only records for individuals satisfying the
sex requirement
}
\description{
Restrict cohort on sex
}
\examples{
library(DrugUtilisation)
library(CohortConstructor)
cdm <- mockDrugUtilisation(numberIndividuals = 100)
cdm$cohort1 \%>\%
  requireSex(sex = "Female")
}
