% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireTableIntersectFlag.R
\name{requireTableIntersectFlag}
\alias{requireTableIntersectFlag}
\title{Require cohort subjects are present in another cohort}
\usage{
requireTableIntersectFlag(
  x,
  targetTable,
  indexDate = "cohort_start_date",
  targetStartDate = startDateColumn(tableName),
  targetEndDate = endDateColumn(tableName),
  censorDate = NULL,
  window = list(c(0, Inf)),
  negate = FALSE,
  name = omopgenerics::tableName(x)
)
}
\arguments{
\item{x}{Cohort table.}

\item{targetTable}{Name of the table that we want to check for intersect.}

\item{indexDate}{Variable in x that contains the date to compute the
intersection.}

\item{targetStartDate}{Date of reference in cohort table, either for start
(in overlap) or on its own (for incidence).}

\item{targetEndDate}{Date of reference in cohort table, either for end
(overlap) or NULL (if incidence).}

\item{censorDate}{Whether to censor overlap events at a specific date or a
column date of x.}

\item{window}{Window to consider events over.}

\item{negate}{If set as TRUE, criteria will be applied as exclusion
rather than inclusion (i.e. require absence in another cohort).}
}
\value{
Cohort table with only those in the other cohort kept
}
\description{
Require cohort subjects are present in another cohort
}
\examples{
library(DrugUtilisation)
library(CohortConstructor)
cdm <- mockDrugUtilisation(numberIndividuals = 100)
cdm$cohort1 \%>\%
  requireCohortIntersectFlag(targetCohortTable = "cohort2",
                             targetCohortId = 1,
                             indexDate = "cohort_start_date",
                             window = c(-Inf, 0))
}
