% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifyCohorts.R
\name{stratifyCohorts}
\alias{stratifyCohorts}
\title{Crate a new cohort table with the specified stratifications.}
\usage{
stratifyCohorts(
  cohort,
  cohortId = NULL,
  strata = list(),
  removeStrata = TRUE,
  name = tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{Vector of cohort definition ids to include. If NULL, all
cohort definition ids will be used.}

\item{strata}{A strata list that point to columns in cohort table.}

\item{removeStrata}{Whether to remove strata columns from final cohort table.}

\item{name}{Name of the new cohort.}
}
\value{
Cohort table stratified.
}
\description{
Crate a new cohort table with the specified stratifications.
}
\examples{
library(CohortConstructor)
library(PatientProfiles)

# cdm <- mockCohortConstructor()
#
# cdm$my_cohort <- conceptCohort(
#   cdm = cdm,
#   conceptSet = list(x = c(1, 2), y = c(2, 7, 8)),
#   name = "my_cohort"
# ) |>
#  addAge(ageGroups = list("child" = c(0, 17), "adult" = c(18, Inf))) |>
#  addSex() |>
#  stratifyCohorts(
#    strata = list("sex, c("sex", "age_group")), name = "my_cohort"
#  )
#
# cdm$mycohort
#
# settings(cdm$my_cohort)
#
# attrition(cdm$my_cohort)

}
