% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCohorts.R
\name{subsetCohorts}
\alias{subsetCohorts}
\title{Generate a cohort table using a subset of cohorts from another table.}
\usage{
subsetCohorts(
  cohort,
  cohortId,
  name = tableName(cohort),
  .softValidation = FALSE
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{Vector of cohort definition ids to include. If NULL all
cohort will be selected.}

\item{name}{Name of the new cohort with the demographic requirements.}

\item{.softValidation}{Whether to perform a soft validation of consistency.
If set to FALSE four additional checks will be performed: 1) check that
cohort end date is not before cohort start date, 2) check that there are no
missing values in required columns, 3) check that cohort duration is all
within observation period, and 4) check that there are no overlapping cohort
entries.}
}
\value{
Cohort table with only cohorts in cohortId.
}
\description{
Generate a cohort table using a subset of cohorts from another table.
}
\examples{
library(CohortConstructor)

cdm <- mockCohortConstructor(nPerson = 100)

cdm$cohort1 |> subsetCohorts(cohortId = 1)

}
