% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireDateRange.R
\name{trimToDateRange}
\alias{trimToDateRange}
\title{Trim cohort dates to be within a date range}
\usage{
trimToDateRange(
  cohort,
  dateRange,
  cohortId = NULL,
  startDate = "cohort_start_date",
  endDate = "cohort_end_date",
  name = tableName(cohort),
  .softValidation = FALSE
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{dateRange}{A window of time during which the index date must have
been observed.}

\item{cohortId}{Vector of cohort definition ids to include. If NULL, all
cohort definition ids will be used.}

\item{startDate}{Variable with earliest date.}

\item{endDate}{Variable with latest date.}

\item{name}{Name of the new cohort with the restriction.}

\item{.softValidation}{Whether to perform a soft validation of consistency.
If set to FALSE four additional checks will be performed: 1) check that
cohort end date is not before cohort start date, 2) check that there are no
missing values in required columns, 3) check that cohort duration is all
within observation period, and 4) check that there are no overlapping cohort
entries.}
}
\value{
The cohort table with record timings updated to only be within the
date range. Any records with all time outside of the range will have
been dropped.
}
\description{
Trim cohort dates to be within a date range
}
\examples{
library(CohortConstructor)
cdm <- mockCohortConstructor()
cdm$cohort1 |>
  trimToDateRange(startDate = "cohort_start_date",
                  endDate = "cohort_end_date",
                  dateRange = as.Date(c("2015-01-01",
                                        "2015-12-31")))
}
