% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionCohorts.R
\name{unionCohorts}
\alias{unionCohorts}
\title{Generate cohort from the union of different cohorts}
\usage{
unionCohorts(
  cohort,
  cohortId = NULL,
  gap = 0,
  cohortName = NULL,
  name = tableName(cohort),
  .softValidation = FALSE
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{Vector of cohort definition ids to include. If NULL, all
cohort definition ids will be used.}

\item{gap}{Number of days between two subsequent cohort entries of a subject
that will be merged in a single cohort entry}

\item{cohortName}{Name of the returned cohort. If NULL, the cohort name will
be created by collapsing the individual cohort names, separated by "_".}

\item{name}{Name of the new cohort table.}

\item{.softValidation}{Whether to perform a soft validation of consistency.
If set to FALSE four additional checks will be performed: 1) check that
cohort end date is not before cohort start date, 2) check that there are no
missing values in required columns, 3) check that cohort duration is all
within observation period, and 4) check that there are no overlapping cohort
entries.}
}
\value{
A cohort table.
}
\description{
Generate cohort from the union of different cohorts
}
\examples{
\donttest{
library(CohortConstructor)

cdm <- mockCohortConstructor(nPerson = 100)

cdm$cohort2 <- cdm$cohort2 |> unionCohorts()
settings(cdm$cohort2)

}
}
