% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearCohorts.R
\name{yearCohorts}
\alias{yearCohorts}
\title{Generate a new cohort_table object restrincting observation to certain years.}
\usage{
yearCohorts(cohort, cohortId = NULL, years, name = tableName(cohort))
}
\arguments{
\item{cohort}{A cohort_table in a cdm_reference.}

\item{cohortId}{Cohort definition id to use. If NULL all the cohort
definition ids in settings will be used.}

\item{years}{Numeric vector of years to use to restrict observation to.}

\item{name}{Name of the new cohort table.}
}
\value{
A cohort table.
}
\description{
Generate a new cohort_table object restrincting observation to certain years.
}
\examples{
library(omock)

cdm <- mockCdmReference() |>
  mockPerson() |>
  mockObservationPeriod() |>
  mockCohort()
cdm$year_restricted <- cdm$cohort |>
  yearCohorts(years = 2010:2019, name = "year_restricted")
cdm$year_restricted
cdm$year_restricted |> settings()
cdm$year_restricted |> attrition()

}
