% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearCohorts.R
\name{yearCohorts}
\alias{yearCohorts}
\title{Generate a new cohort table restricting cohort entries to certain years}
\usage{
yearCohorts(cohort, years, cohortId = NULL, name = tableName(cohort))
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{years}{Numeric vector of years to use to restrict observation to.}

\item{cohortId}{IDs of the cohorts to modify. If NULL, all cohorts will be
used; otherwise, only the specified cohorts will be modified, and the
rest will remain unchanged.}

\item{name}{Name of the new cohort table.}
}
\value{
A cohort table.
}
\description{
Generate a new cohort table restricting cohort entries to certain years
}
\examples{
library(CohortConstructor)

cdm <- mockCohortConstructor(nPerson = 100)

cdm$cohort1 <- cdm$cohort1 |> yearCohorts(years = 2000:2002)
settings(cdm$cohort1)

}
